python/requirements.txt
fastapi
uvicorn[standard]
aiosqlite
PyJWT
httpx
python-dotenv
